+++ docs/userguide/docs/changes.md
@@ -31,6 +31,7 @@ createSimulation {
 ComponentGenerator(uniform(0,1)){ counter -> "smthg no$counter"}
 ```
 
+* Changed scheduling priority from `Int` to inline class `Priority` (with defaults `NORMAL`, `HIGH`, `LOW`) in all interaction methods for more typesafe API.
 
 * New Example: ["The ferryman"](examples/ferryman.md)
 * New Example: [Office Tower](examples/office_tower.md)